requirements.txt
apache-airflow-providers-mongo==3.5.0
apache-airflow-providers-docker==3.9.1
apache-airflow-providers-papermill==3.6.0
jupyterlab==4.0.11
ipykernel==6.29.0
neo4j==5.16.0 
docker==6.1.3 
